+++ concierge-client/src/test/java/com/grookage/concierge/client/ConciergeClientTest.java
@@ -6,8 +6,8 @@ import com.grookage.concierge.client.refresher.ConciergeClientRefresher;
 import com.grookage.concierge.client.refresher.ConciergeClientSupplier;
 import com.grookage.concierge.client.serde.SerDe;
 import com.grookage.concierge.client.serde.SerDeFactory;
-import com.grookage.concierge.models.NamespaceRequest;
 import com.grookage.concierge.models.ResourceHelper;
+import com.grookage.concierge.models.SearchRequest;
 import com.grookage.concierge.models.config.ConfigKey;
 import com.grookage.concierge.models.ingestion.ConfigurationResponse;
 import com.grookage.leia.provider.config.LeiaHttpConfiguration;
@@ -29,7 +29,7 @@ class ConciergeClientTest {
     @Test
     @SneakyThrows
     void testConciergeClient(WireMockRuntimeInfo wireMockRuntimeInfo) {
-        final var namespaceRequest = NamespaceRequest.builder()
+        final var namespaceRequest = SearchRequest.builder()
                 .namespaces(Set.of("concierge"))
                 .build();
         final var configResponse = ResourceHelper.getResource("configurationResponse.json",