+++ concierge-dw/src/main/java/com/grookage/conceirge/dwserver/resources/ConfigResource.java
@@ -19,9 +19,8 @@ package com.grookage.conceirge.dwserver.resources;
 import com.codahale.metrics.annotation.ExceptionMetered;
 import com.codahale.metrics.annotation.Timed;
 import com.grookage.concierge.core.services.ConfigService;
-import com.grookage.concierge.models.ConfigNamesRequest;
 import com.grookage.concierge.models.MapperUtils;
-import com.grookage.concierge.models.NamespaceRequest;
+import com.grookage.concierge.models.SearchRequest;
 import com.grookage.concierge.models.config.ConciergeRequestContext;
 import com.grookage.concierge.models.config.ConfigDetails;
 import com.grookage.concierge.models.config.ConfigKey;
@@ -35,7 +34,6 @@ import lombok.extern.slf4j.Slf4j;
 import javax.annotation.security.PermitAll;
 import javax.inject.Singleton;
 import javax.validation.Valid;
-import javax.validation.constraints.NotEmpty;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import java.util.ArrayList;
@@ -91,9 +89,9 @@ public class ConfigResource {
     @ExceptionMetered
     @Path("/details/active")
     public List<ConfigurationResponse> getActiveConfigs(
-            @QueryParam("ignoreCache") boolean ignoreCache, @Valid final NamespaceRequest namespaceRequest
+            @QueryParam("ignoreCache") boolean ignoreCache, @Valid final SearchRequest searchRequest
     ) {
-        return getConfigurationResponses(configService.getActiveConfigs(toRequestContext(ignoreCache), namespaceRequest.getNamespaces()));
+        return getConfigurationResponses(configService.getActiveConfigs(toRequestContext(ignoreCache), searchRequest.getNamespaces()));
     }
 
     @POST
@@ -105,7 +103,7 @@ public class ConfigResource {
             @PathParam("namespace") final String namesapce,
             @PathParam("configName") final String configName
     ) {
-        final var config = configService.getLatestActiveConfig(toRequestContext(ignoreCache),namesapce, configName).orElse(null);
+        final var config = configService.getLatestActiveConfig(toRequestContext(ignoreCache), namesapce, configName).orElse(null);
         return getConfigurationResponses(null == config ? List.of() : List.of(config));
     }
 
@@ -114,22 +112,9 @@ public class ConfigResource {
     @ExceptionMetered
     @Path("/details")
     public List<ConfigurationResponse> getConfigs(
-            @QueryParam("ignoreCache") boolean ignoreCache, @Valid final NamespaceRequest namespaceRequest) {
-        return getConfigurationResponses(configService.getConfigs(toRequestContext(ignoreCache), namespaceRequest.getNamespaces()));
+            @QueryParam("ignoreCache") boolean ignoreCache, @Valid final SearchRequest searchRequest) {
+        return getConfigurationResponses(configService.getConfigs(toRequestContext(ignoreCache), searchRequest));
     }
-
-    @POST
-    @Timed
-    @ExceptionMetered
-    @Path("/{namespace}")
-    public List<ConfigurationResponse> getConfigs(
-            @QueryParam("ignoreCache") boolean ignoreCache,
-            @PathParam("namespace") @NotEmpty final String namespace,
-            @Valid final ConfigNamesRequest configNamesRequest) {
-        return getConfigurationResponses(configService.getConfigs(toRequestContext(ignoreCache),
-                namespace, configNamesRequest.getConfigNames()));
-    }
-
 
     @GET
     @Timed