+++ concierge-models/src/main/java/com/grookage/concierge/models/config/ConfigDetails.java
@@ -45,4 +45,5 @@ public class ConfigDetails implements Comparable<ConfigDetails> {
     public int compareTo(ConfigDetails o) {
         return configKey.compareTo(o.getConfigKey());
     }
+
 }